pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pybites-alarm"
authors = [{name = "PyBites", email = "support@pybit.es"}]
readme = "README.md"
requires-python = ">=3.7"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "python-dotenv >=0.19.0",
    "pydub >=0.25.1",
    "simpleaudio >=1.0.4",
    "audio-program-generator >=1.6.3",
]

[tool.flit.module]
name = "alarm"

[project.urls]
Source = "https://github.com/PyBites-Open-Source/pybites-alarm"

[project.scripts]
alarm = "alarm.__main__:main"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = python3.7,py38,py39
isolated_build = True

[testenv]
deps =
    pytest
    psutil==5.8.0
    SpeechRecognition==3.8.1
commands = pytest {posargs}
"""
